begin adding step functions for orchestration
