+++ .gitignore
@@ -87,3 +87,6 @@ target/
 
 # Mypy cache
 .mypy_cache/
+
+# any env files
+*.env.*
