+++ src/lambda/pet_match_stack/get_auth/requirements.txt
@@ -1,0 +1,2 @@
+requests==2.28.1
+python-dotenv==0.21.0
