--- src/lambda/pet_match_stack/get_data/requirements.txt
@@ -1,1 +1,0 @@
-requests==2.28.1
