+++ src/lambda/pet_match_stack/template.yaml
@@ -26,10 +26,10 @@ Resources:
             Path: /hello
             Method: get
 
-  GetDataFunction:
+  GetAuthFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: get_data/
+      CodeUri: get_auth/
       Handler: app.lambda_handler
       Runtime: python3.9
       Architectures:
@@ -38,7 +38,7 @@ Resources:
         GetData:
           Type: Api # More info about API Event Source: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#api
           Properties:
-            Path: /get_data
+            Path: /get_auth
             Method: get
 
   # Base level layer with common Python modules
@@ -63,18 +63,14 @@ Outputs:
   HelloWorldFunction:
     Description: "Hello World Lambda Function ARN"
     Value: !GetAtt HelloWorldFunction.Arn
-  
-  # HelloWorldFunctionIamRole:
-  #   Description: "Implicit IAM Role created for Hello World function"
-  #   Value: !GetAtt HelloWorldFunctionRole.Arn
 
   PetMatchApi:
     Description: "API Gateway endpoint URL for Dev stage for PetMatch Get Data function"
     Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Dev/get_data/"
 
-  GetDataFunction:
+  GetAuthFunction:
     Description: "Get Data Lambda Function ARN"
-    Value: !GetAtt GetDataFunction.Arn
+    Value: !GetAtt GetAuthFunction.Arn
   
   PetMatchLambdaRole:
     Description: "Implicit IAM Role created for Lambdas in the Pet Match Stack"
