+++ src/step_functions/aws-stepfunctions-local-credentials.txt
@@ -1,0 +1,28 @@
+### AWS VARS ###
+
+AWS_DEFAULT_REGION=us-east-1
+AWS_ACCESS_KEY_ID=demo
+AWS_SECRET_ACCESS_KEY=demo
+
+### AWS VARS ###
+
+### STACK VARS ###
+
+STEP_FUNCTIONS_ENDPOINT=host.docker.internal
+WAIT_TIME_SCALE=60
+LAMBDA_ENDPOINT=host.docker.internal.lambda
+
+### STACK VARS ###
+
+
+### OPTIONS ###
+
+BATCH_ENDPOINT=VALUE
+DYNAMODB_ENDPOINT=VALUE
+ECS_ENDPOINT=VALUE
+GLUE_ENDPOINT=VALUE
+SAGE_MAKER_ENDPOINT=VALUE
+SQS_ENDPOINT=VALUE
+SNS_ENDPOINT=VALUE
+
+### OPTIONS ###
